+++ Editor/Enums/EPropertyType.cs
@@ -1,0 +1,10 @@
+﻿namespace Unity_Free_Online_Config.Editor.Enums
+{
+    public enum EPropertyType
+    {
+        String,
+        Int,
+        Float,
+        Double
+    }
+}