+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [v0.2.0]
+- "Create new config" editor window.
+- Basic code generation
+
 # [v0.1.1]
 - Naming refactor.
 - README.md guideline from spreadsheet creation to results.